pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rexsyn-nexus-light"
version = "0.4.0"
description = "Light OSS demo of the RExSyn Nexus backend (placeholder science + CTA for full edition)"
authors = [{name = "YUN", email = "yun@flamehaven.ai"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110,<1.0",
    "uvicorn>=0.23,<1.0",
    "pydantic>=2.6,<3.0",
    "email-validator>=2.2,<3.0",
    "numpy>=1.26,<2.0",
    "sqlalchemy>=2.0,<3.0",
    "python-multipart>=0.0.9",
    "passlib[bcrypt]>=1.7,<2.0",
    "python-jose>=3.3,<4.0",
    "prometheus-client>=0.19,<1.0",
    "redis>=5.0,<6.0",
    "slowapi>=0.1.9"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "httpx>=0.25",
    "ruff>=0.6",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "--cov=backend/app --cov-report=term-missing -q"

[tool.ruff]
line-length = 120
target-version = "py311"
extend-exclude = [
    "meta_pytest",
    "tests",
    "monitoring",
    "src",
    "tools",
    "workers",
    "docs",
    "backend/app/services/report_generator.py",
    "backend/app/services/risk_service.py",
    "backend/app/services/md_refinement.py",
    "backend/app/tasks/prediction_tasks.py",
]

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["E501"]
